+++ src/core/weechat.c
@@ -187,6 +187,7 @@ weechat_display_usage ()
           "directory and delete it on exit\n"
           "                           (incompatible with option \"-d\")\n"
           "  -h, --help               display this help and exit\n"
+          "  -i, --build-info         display build information and exit\n"
           "  -l, --license            display WeeChat license and exit\n"
           "  -p, --no-plugin          don't load any plugin at startup\n"
           "  -P, --plugins <plugins>  load only these plugins at startup\n"
@@ -258,6 +259,7 @@ weechat_parse_args (int argc, char *argv[])
         { "dir",         required_argument, NULL, 'd'               },
         { "temp-dir",    no_argument,       NULL, 't'               },
         { "help",        no_argument,       NULL, 'h'               },
+        { "build-info",  no_argument,       NULL, 'i'               },
         { "license",     no_argument,       NULL, 'l'               },
         { "no-plugin",   no_argument,       NULL, 'p'               },
         { "plugins",     required_argument, NULL, 'P'               },
@@ -287,7 +289,7 @@ weechat_parse_args (int argc, char *argv[])
     opterr = 0;
 
     while ((opt = getopt_long (argc, argv,
-                               ":acd:thlpP:r:sv",
+                               ":acd:thilpP:r:sv",
                                long_options, NULL)) != -1)
     {
         switch (opt)
@@ -314,6 +316,10 @@ weechat_parse_args (int argc, char *argv[])
                 break;
             case 'h': /* -h / --help */
                 weechat_display_usage ();
+                weechat_shutdown (EXIT_SUCCESS, 0);
+                break;
+            case 'i': /* -i / --build-info */
+                debug_build_info ();
                 weechat_shutdown (EXIT_SUCCESS, 0);
                 break;
             case 'l': /* -l / --license */
@@ -601,11 +607,11 @@ void
 weechat_init_gettext ()
 {
     weechat_locale_ok = (setlocale (LC_ALL, "") != NULL);   /* init gettext */
-#ifdef ENABLE_NLS
+#if ENABLE_NLS == 1
     bindtextdomain (PACKAGE, LOCALEDIR);
     bind_textdomain_codeset (PACKAGE, "UTF-8");
     textdomain (PACKAGE);
-#endif /* ENABLE_NLS */
+#endif /* ENABLE_NLS == 1 */
 
 #ifdef HAVE_LANGINFO_CODESET
     weechat_local_charset = strdup (nl_langinfo (CODESET));